+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for plutarch
 
+# Unreleased
+
+- `TermCont`: Parametrize by result type; add `MonadFail` instance; etc.
+
+  Also, export from `Plutarch.TermCont`, and then from `Plutarch.Prelude` (TermCont is no longer exported by `Plutarch.Internal`).
+
+  Started by: [#226](https://github.com/Plutonomicon/plutarch/pull/226)
+
 # 1.1.0
 
 - General repository changes.